backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
channels==4.1.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
daphne==4.1.0
requests==2.31.0
Pillow==10.3.0
pytest==8.1.1
pytest-django==4.7.0
pytest-asyncio==0.23.6
